+++ src/app/locales/en/components/file-browser/fb-info-modal.js
@@ -10,32 +10,54 @@ export default {
   owner: 'Owner',
   close: 'Close',
   size: 'Size',
-  rest: 'Public REST endpoint',
-  restTag: 'REST',
-  restTipIntro: 'The Onezone\'s public REST API can be used to access information and contents of all shared files and directories, without any authentication. It redirects to the corresponding REST API in one of the supporting Oneproviders. The Oneprovider is chosen dynamically and may change in time, so the redirection URL should not be cached.',
-  restTipLinkName: 'REST API',
-  restTipSpecificIntro: 'This endpoint returns {{typeDescription}}.',
-  restTipSpecificType: {
-    listSharedDirectoryChildren: 'the list of directory files and subdirectories',
-    downloadSharedFileContent: {
-      file: 'the binary file content',
-      dir: 'a compressed (GZIP) TAR archive with directory contents',
-    },
-    getSharedFileAttributes: 'basic attributes of a file or directory',
-    getSharedFileJsonMetadata: 'custom JSON metadata associated with a file or directory',
-    getSharedFileRdfMetadata: 'custom RDF metadata associated with a file or directory',
-    getSharedFileExtendedAttributes: 'custom extended attributes (xattrs) associated with a file or directory',
+  api: 'Public API',
+  apiCommandTipIntro: {
+    rest: 'The Onezone\'s public REST API can be used to access information and contents of all shared files and directories, without any authentication. It redirects to the corresponding REST API in one of the supporting Oneproviders. The Oneprovider is chosen dynamically and may change in time, so the redirection URL should not be cached.',
+    xrootd: 'This environment offers an XRootD server that exposes all Open Data collections (shared files and directories that had been registered under a handle) for public, read-only access. Below are some basic XRootD commands that can be used to browse and download the data, assuming that the XRootD CLI tools are already installed.',
   },
-  restUrlType: {
-    listSharedDirectoryChildren: 'List directory files and subdirectories',
-    downloadSharedFileContent: {
-      file: 'Download file content',
-      dir: 'Download directory (tar.gz)',
+  apiCommandTipLinkName: {
+    rest: 'REST API',
+    xrootd: 'XRootD',
+  },
+  apiCommandTipSpecificIntro: {
+    rest: 'This endpoint returns {{typeDescription}}.',
+    xrootd: 'This shell command {{typeDescription}}.',
+  },
+  apiCommandTipSpecificType: {
+    rest: {
+      listSharedDirectoryChildren: 'the list of directory files and subdirectories',
+      downloadSharedFileContent: 'the binary file content',
+      downloadSharedDirectoryContent: 'a compressed (GZIP) TAR archive with directory contents',
+      getSharedFileAttributes: 'basic attributes of a file or directory',
+      getSharedFileJsonMetadata: 'custom JSON metadata associated with a file or directory',
+      getSharedFileRdfMetadata: 'custom RDF metadata associated with a file or directory',
+      getSharedFileExtendedAttributes: 'custom extended attributes (xattrs) associated with a file or directory',
     },
-    getSharedFileAttributes: 'Get attributes',
-    getSharedFileJsonMetadata: 'Get JSON metadata',
-    getSharedFileRdfMetadata: 'Get RDF metadata',
-    getSharedFileExtendedAttributes: 'Get extended attributes (xattrs)',
+    xrootd: {
+      listSharedDirectoryChildren: 'returns the list of directory files and subdirectories',
+      downloadSharedFileContent: 'downloads the file',
+      downloadSharedDirectoryContent: 'downloads the directory recursively',
+    },
+  },
+  apiCommandTitle: {
+    rest: {
+      listSharedDirectoryChildren: 'List directory files and subdirectories',
+      downloadSharedFileContent: 'Download file content',
+      downloadSharedDirectoryContent: 'Download directory (tar.gz)',
+      getSharedFileAttributes: 'Get attributes',
+      getSharedFileJsonMetadata: 'Get JSON metadata',
+      getSharedFileRdfMetadata: 'Get RDF metadata',
+      getSharedFileExtendedAttributes: 'Get extended attributes (xattrs)',
+    },
+    xrootd: {
+      listSharedDirectoryChildren: 'List directory files and subdirectories',
+      downloadSharedFileContent: 'Download file',
+      downloadSharedDirectoryContent: 'Download directory (recursively)',
+    },
+  },
+  apiType: {
+    rest: 'REST',
+    xrootd: 'XRootD',
   },
   fileType: {
     file: 'file',