+++ src/tests/unit/services/rest-generator-test.js
@@ -4,11 +4,12 @@ import { setupTest } from 'ember-mocha';
 import { registerService, lookupService } from '../../helpers/stub-service';
 import { set } from '@ember/object';
 import Service from '@ember/service';
+import { sharedRestFileTemplate } from 'oneprovider-gui/services/mocks/onedata-connection';
 
 const OnedataConnection = Service.extend({
   init() {
     this._super(...arguments);
-    this.set('restTemplates', {});
+    this.set('apiTemplates', { rest: {} });
   },
 });
 
@@ -34,18 +35,25 @@ describe('Unit | Service | rest generator', function () {
     'getSharedFileRdfMetadata',
     'getSharedFileExtendedAttributes',
   ].forEach(methodName => testShareDataCurlGeneratingMethod(methodName));
+
+  // download directory content uses the same template as download file
+  testShareDataCurlGeneratingMethod(
+    'downloadSharedDirectoryContent',
+    'downloadSharedFileContent'
+  );
 });
 
 function testShareDataCurlGeneratingMethod(methodName, templateName = methodName) {
-  it(`generates URL for ${methodName} method using "${templateName}" template`, function () {
+  it(`generates curl command for ${methodName} method using "${templateName}" template`, function () {
     const id = '1234';
     set(
       lookupService(this, 'onedataConnection'),
-      `restTemplates.${templateName}`,
-      `https://onezone.org/{{id}}/${methodName}`
+      `apiTemplates.rest.${templateName}`,
+      sharedRestFileTemplate(methodName)
     );
     const service = this.subject();
-    expect(service[methodName](id))
-      .to.equal(`curl -L https://onezone.org/1234/${methodName}`);
+    expect(service[methodName]({ cdmiObjectId: id })).to.equal(
+      `curl -L 'https://test.onedata.org/api/v3/onezone/shares/data/1234/${methodName}'`
+    );
   });
 }